move remaining validated structs to controller
